chore: Renames helpers to utilities
