+++ apps/site/pages/components/atoms/slider.mdx
@@ -11,7 +11,7 @@ import { OverviewSection } from 'site/components/OverviewSection'
 import { SectionItem, SectionList } from 'site/components/SectionItem'
 import { Slider, Price, PriceRange } from '@faststore/ui'
 import PriceRangeUsage from 'site/components/PriceRange/PriceRangeUsage'
-import { useFormattedPrice } from 'site/components/Helpers/usePriceFormatter'
+import { useFormattedPrice } from 'site/components/utilities/usePriceFormatter'
 
 <header>
