+++ apps/site/pages/components/molecules/gift.mdx
@@ -13,7 +13,7 @@ import { SectionItem, SectionList } from 'site/components/SectionItem'
 import { Badge, Gift, GiftContent, GiftImage, Icon, Price } from '@faststore/ui'
 import { TagIcon, ShoppingCart } from '@faststore/components'
 
-import { useFormattedPrice } from 'site/components/Helpers/usePriceFormatter'
+import { useFormattedPrice } from 'site/components/utilities/usePriceFormatter'
 
 <header>
