+++ apps/site/pages/components/molecules/table.mdx
@@ -19,7 +19,7 @@ import {
   Price,
 } from '@faststore/ui'
 import { Ruler, ShoppingCart } from '@faststore/components'
-import { useFormattedPrice } from 'site/components/Helpers/usePriceFormatter'
+import { useFormattedPrice } from 'site/components/utilities/usePriceFormatter'
 
 <header>
